+++ MANIFEST.in
@@ -1,4 +1,4 @@
-recursive-include rf_doc_server *.css
-recursive-include rf_doc_server *.png
-recursive-include rf_doc_server *.html
-recursive-include rf_doc_server *.py
+recursive-include rfdocserver *.css
+recursive-include rfdocserver *.png
+recursive-include rfdocserver *.html
+recursive-include rfdocserver *.py
