+++ readme.md
@@ -1,16 +1,16 @@
-# rf-doc-server
+# rfdocserver
 
 This python modules goal is to make browsing robot framework keyword documentation easier by generating browseable documentation with robot framework libdoc and host it with flask web server.
 
 ## Installation
 
-You will need to install some python3 dependencies for this project to work. This can be done with [pip](https://docs.python.org/3/installing/index.html).
+You can install this package with pip:
 
-- `pip3 install -r requirements.txt`
+`pip3 install rfdocserver`
 
 
 ## Running
 
 This project should be run as python3 module.
 
-`python3 -m rf-doc-server <path to robot file root>`
+`python3 -m rfdocserver <path to robot file root>`
